Allow unprocessed external links in `img` tags
Why:
By default, all `<img>` tags are processed, and even if the docs clearly
stated that you'd need to turn off `optimizeAll` to use any external
sources in your regular image tags, it would be a shame to not get the
benefits on all the rest of your internal images.

How:
**Add a check for external paths on all images**
This is done simply using a RegExp

**Return an object from `getPathname`**
The previous strategy for preventing image processing was throwing an
error. Since we are now expecting some images to be unprocessable, it is
no longer an error. As such, `getPathname` now always returns an object
with all the information necessary to determine if a path should be
processed further.

**Change error strategy inside `replaceInImg` and `replaceInComponent`**
We are now expecting to catch images that are unprocessable before
attempting to optimize. This means that any strange cases are actually
unforseen errors that we can catch in a new try/catch block inside the
`replaceInImg` function. Since developers will likely be using the Image
component in specific cases, the `replaceInComponent` will log expected
path issues to the console, letting everything else attempt to run as
is.