Fix uncaught error when no value for src
In removing the try/catch from `getPathname` I uncovered another area
that was benfitting from the subsequent error handler: getProp.

How:
**Add value checking before assuming a src**
`getProp(node, 'src')` regularly failed, but was caught and ignored.
Now, I have added a `getSrc` fn that expects a `src` may not exist. It
returns a valid, but empty, object in those cases.

Additionally, `getProp` --- currently only used by `getSrc` --- checks
for a prop before trying to return its value.

Previous attempts to get the src of an image will be fine. The one that
was failing (inside `getPathname`) will be covered by the test for
`value.data`.