Fix resizing bug
How:
**Remove null sizes from sizes array in `replaceInComponent`**
I've moved the entire loop of resizing out of the calling function and
filtered all null values in a new function called `createSizes`. (Null
values are produced when the original image is smaller than the desired
resized image.)

**Pass original size to `resize` if smallest size is bigger than image**
If an image referenced by the Image component was smaller than the
smallest image size found in user-created options, then no images would
be created. If this is the case, we just pass in the original image's
size.