Add feature: allow initial slash for images
Why:
All images that can be used inside the application will have to be
resolved from the static directory. As such, the proper way to resolve
those images inside HTML would actually be from the root of the site,
found at "/".

How:
Remove a single leading slash from the front of any image if it is an
internal image.