Catch up changelog & version number
The version number and changelog were a bit out of sync with actual
releases on NPM.

Here, I have updated the version to the _next_ smallest possible version
and appended the `-pre` identifier to signify that this is a prerelease
state of the codebase. At the time of a new release the actual bump size
will be determined and that can be published to NPM.

I have also updated the formatting on the changelog.