Add Changelog tooling
Why:
This should make the process of updating version numbers and changelog
entries easier and less error prone.

How:
This is a bit of code I have worked on in the past, with a couple
refinements for this repo. When you run `yarn release`, the script will
take over and help you to tag a release with a proper, new version
number.

First, it prompts you to choose a release type: patch, minor, or major.

Next, it updates the changelog and package.json with the release date
and new version number. It creates a commit with these changes and tags
it.

Then, it waits while you publish to NPM in a separate terminal. Once you
have indicated you are done, it will create a new commit with a
continuing version number placeholder in package.json.

Finally, it will push your two new commits and your new tag up to your
origin repo.