Add feature: process image folders
By adding a few options to the config object, we can process all the
image files in a given set of folders (the first time the parser runs,
only) without the need to read them as a literal string in a template.
This will allow developers to target optimized versions of images in
regular `<img>` tags inside their templates using logic that we cannot
parse at build time.

This is both a stop-gap to actually parsing strings to get path values,
and also a useful feature in its own right.

How:

**Prevent inadvertent mutation of default settings**
This was causing some tests to fail.

**Add folder processing feature**
See the new docs and tests for a full understanding.