Fix "process folders once"
This fixes one bug where we are doing more work than necessary and one
potentiality where we might miss some intended images.

The bug is that when `getPreprocessor` is called multiple times (which
is possible in some use cases), the old `ran` boolean was reset to
`false`, causing us to attempt to re-process all the images in the
folders chosen for `processFolders`. Ultimately, we'd not re-process any
images that already exist in the target location, but we were still
doing more work than necessary. Moving the check for having already run
into module scope clears that up.

However, it is possible that separate calls to `getPreprocessor`
could potentially pass in *different options*. So we cannot just use a
simple boolean to determine if we should be attempting to process any
folders. If options change to process an entirely different folder, we
would miss it.

So to determine if we have already processed the folders for the current
set of options, we save a stringified version of the relevant options
and check if that string has already been created or not.